summit: clean up packages.yaml, add extended
packages-extended.yaml is contains more externals, many
generated using the spack-gen-external-packages.py script.
This version is not as well tested but can save a lot of time
building. It can serve as an example to pick and choose items
to add to packages.yaml if they fail to build cleanly from spack.