spack: don't include effis in wdmapp meta package
The "useful" EFFIS needs to be built with python=type=full, so that
job composition/submission etc is available.

Unfortunately, it is not possible to do this variant in the metapackage,
since it requires python3, but xgc requires python2 via petsc, and spack
does not allow for that.

Instead the defaults for EFFIS are set in packages.yaml so that a separate
`spack install effis` will do the right thing.