doc: use git urls, promote info for finding paths
The testcases app is private, and the https url does not use keys and
prompts for github password. The wdmapp-config repo is currently
public, so the https url still works, but I think it's better to be
consistent and it may become private.

The instructions for finding the paths for coupler/gene/xgc are
important, and notes are more for less important details and asides.